+++ bw2data/__init__.py
@@ -33,7 +33,7 @@ __all__ = [
     "weightings",
 ]
 
-__version__ = (4, 0, "dev53")
+__version__ = (4, 0, "dev54")
 
 from bw2data.configuration import config, labels
 from bw2data.project import projects