+++ CHANGES.md
@@ -1,6 +1,6 @@
 # `bw2data` Changelog
 
-## DEV
+## 4.0.DEV54 (2024-09-14)
 
 * Allow `get_node(key=(<some tuple))`
 * Raise `UnknownObject` when writing LCIA methods referring to non-existent flows